replit_mcp_server.py
import sys
from mcp.server import Server

# Tool functions


def echo_tool(input: str) -> str:
    return f"Echo: {input}"


def shout_tool(input: str) -> str:
    return f"🔥 {input.upper()} 🔥"


def reverse_tool(input: str) -> str:
    return f"↩️ {input[::-1]}"


# Create the server
server = Server(name="replit-mcp")

# Tool dispatcher: this is how we handle multiple tools


def handle_tool(name: str, args: dict):
    if name == "echo":
        return echo_tool(args.get("input", ""))
    elif name == "shout":
        return shout_tool(args.get("input", ""))
    elif name == "reverse":
        return reverse_tool(args.get("input", ""))
    else:
        return f"❌ Unknown tool: {name}"


# Hook dispatcher into MCP’s tool system
server.call_tool = handle_tool


if __name__ == "__main__":
    print("🚀 MCP Server running (tools: echo, shout, reverse)")
    server.run(
        read_stream=sys.stdin.buffer,
        write_stream=sys.stdout.buffer,
        initialization_options={},
    )

# ==============================================================================
# © 2025 Everett Nathaniel Christman & Misty Gail Christman
# The Christman AI Project — Luma Cognify AI
# All rights reserved. Unauthorized use, replication, or derivative training 
# of this material is prohibited.
# Core Directive: "How can I help you love yourself more?" 
# Autonomy & Alignment Protocol v3.0
# ==============================================================================
