requirements.txt
absl-py==2.3.1
aiofiles==25.1.0
annotated-types==0.7.0
anthropic==0.71.0
anyio==4.11.0
astunparse==1.6.3
attrs==25.4.0
audioread==3.0.1
beautifulsoup4==4.14.2
blinker==1.9.0
boto3==1.40.57
botocore==1.40.57
bs4==0.0.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cryptography==46.0.3
decorator==5.2.1
distro==1.9.0
docstring_parser==0.17.0
exceptiongroup==1.3.0
fastapi==0.119.1
filelock==3.20.0
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
flatbuffers==25.9.23
fsspec==2025.9.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.76.0
gTTS==2.5.4
h11==0.16.0
h5py==3.15.1
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.1
jmespath==1.0.1
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==3.11.3
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
llvmlite==0.45.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
msgpack==1.1.2
namex==0.1.0
networkx==3.4.2
nltk==3.9.2
numba==0.62.1
numpy==1.26.4
ollama==0.6.0
openai==2.6.0
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
platformdirs==4.5.0
pooch==1.8.2
protobuf==4.25.8
pycparser==2.23
pydantic==1.10.15
pydantic_core==2.41.4
pygame==2.6.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
rich==14.2.0
rpds-py==0.28.0
s3transfer==0.14.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.3
soundfile==0.13.1
soupsieve==2.8
soxr==1.0.0
SpeechRecognition==3.14.3
SQLAlchemy==2.0.44
srt==3.5.3
starlette==0.48.0
sympy==1.14.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.67.1
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
vosk==0.3.44
websockets==15.0.1
Werkzeug==3.1.3
wrapt==2.0.0
